Рефакторинг страницы редактирования
